API/requirements.txt
absl-py==1.0.0
appdirs==1.4.4
asgiref==3.5.0
astunparse==1.6.3
audioread==2.1.9
autopep8==1.6.0
boto3==1.24.24
botocore==1.27.24
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
decorator==5.1.1
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.11.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.3
jmespath==1.0.1
joblib==1.1.0
json-encoder==0.4.4
libclang==14.0.1
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.6
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pooch==1.6.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.7.3
singledispatch==3.7.0
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
django-filter==22.1
zipp==3.8.0
tensorflow==2.8.0
keras==2.8.0